requirements.txt
langchain==0.1.13
langchain-community==0.0.29
langchain-google-genai==1.0.1
google-generativeai==0.4.1
pypdf==4.1.0
python-dotenv==1.0.1
streamlit==1.32.2
supabase==2.3.0
cohere==5.0.0
langchain-cohere==0.0.29
docx2txt==0.8
streamlit-authenticator==0.2.3
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3